requirements.txt
joblib==1.1.0
kornia==0.6.4
matplotlib==3.5.1
numpy==1.22.3
openslide_python==1.1.2
pandas==1.4.2
Pillow==9.3.0
PyYAML==6.0
scikit_image==0.19.2
scikit_learn==1.2.0
skimage==0.0
staintools==2.1.2
tensorboardX==2.5.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
